+++ src/main/java/org/pkwmtt/repository/GeneralGroupRepository.java
@@ -3,5 +3,5 @@ package org.pkwmtt.repository;
 import org.pkwmtt.entity.GeneralGroup;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface GeneralGroupRepository extends JpaRepository<GeneralGroup, Long> {
+public interface GeneralGroupRepository extends JpaRepository<GeneralGroup, Integer> {
 }
